Fixed some casting issues in Options

+++ Options.cpp
@@ -53,7 +53,7 @@ const Options* Options::make_options(int argc, const char* argv[])
     if (result.count("time"))
     {
       if(result["time"].as<int>() > 0)
-        opt->time_limit = result["time"].as<unsigned>();
+        opt->time_limit = (unsigned)result["time"].as<int>();
       else
         throw cxxopts::OptionException("Time limit must be positive");
     }
@@ -69,7 +69,7 @@ const Options* Options::make_options(int argc, const char* argv[])
     if(result.count("warmup-samples"))
     {
       if(result["warmup-samples"].as<int>() > 0)
-        opt->warmup_samples = (unsigned)result["warmup-samples"].as<unsigned>();
+        opt->warmup_samples = (unsigned)result["warmup-samples"].as<int>();
       else
         throw cxxopts::OptionException("Number of warmup samples must be positive");
     }
@@ -92,7 +92,7 @@ const Options* Options::make_options(int argc, const char* argv[])
     if(result.count("period-a"))
     {
       if(result["period-a"].as<int>() > 0)
-        opt->pparams_a.period = (unsigned)result["period-a"].as<unsigned>();
+        opt->pparams_a.period = (unsigned)result["period-a"].as<int>();
       else
         throw cxxopts::OptionException("Pruning period cannot be negative, use '-p none' instead");
     }
@@ -100,7 +100,7 @@ const Options* Options::make_options(int argc, const char* argv[])
     if(result.count("period-b"))
     {
       if(result["period-b"].as<int>() > 0)
-        opt->pparams_b.period = (unsigned)result["period-b"].as<unsigned>();
+        opt->pparams_b.period = (unsigned)result["period-b"].as<int>();
       else
         throw cxxopts::OptionException("Pruning period cannot be negative, use '-q none' instead");
     }
